requirements.txt
bcrypt==3.2.2
cffi==1.15.0
click==8.1.3
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.15.0
packaging==21.3
psycopg2-binary==2.9.3
pycparser==2.21
pyparsing==3.0.8
SQLAlchemy==1.4.36
Werkzeug==2.1.2
zipp==3.8.0
pyjwt==2.4.0
